moved mssautoplot to mslib/atuoplot
